Add schema validation

+++ services/health-diary-service/app/database/models.py
@@ -1,5 +1,5 @@
 import uuid
-from sqlalchemy import Column, Integer, Text, DateTime, UUID, Numeric
+from sqlalchemy import Column, Integer, Text, DateTime, UUID, REAL
 
 from .database import Base
 
@@ -10,10 +10,10 @@ class PageDiary(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, index=True)
     id_user = Column(UUID(as_uuid=True), nullable=False)
     pulse = Column(Integer, nullable=False)
-    temperature = Column(Numeric(3, 1), nullable=False)
+    temperature = Column(REAL, nullable=False)
     upper_pressure = Column(Integer, nullable=False)
     lower_pressure = Column(Integer, nullable=False)
     oxygen_level = Column(Integer)
-    sugar_level = Column(Numeric(4, 2))
+    sugar_level = Column(REAL)
     comment = Column(Text)
     create_date = Column(DateTime(timezone=True))
